+++ src/IdFactory.php
@@ -8,14 +8,27 @@ namespace byrokrat\id;
 class IdFactory
 {
     /**
+     * @deprecated Will be removed in version 2. Use createId() instead.
+     */
+    public function create($raw)
+    {
+        trigger_error(
+            'create() is deprecated and will be removed in version 2. Use createId() instead.',
+            E_USER_DEPRECATED
+        );
+
+        return $this->createId($raw);
+    }
+
+    /**
      * Create ID object from raw id string
      *
-     * @param  string $rawId Raw id string
+     * @param  string $raw Raw id string
      * @return IdInterface
-     * @throws Exception\UnableToCreateIdException Always throws exception
+     * @throws Exception\UnableToCreateIdException If unable to create id
      */
-    public function create($rawId)
+    public function createId($raw)
     {
-        throw new Exception\UnableToCreateIdException("Unable to create ID for number '{$rawId}'");
+        throw new Exception\UnableToCreateIdException("Unable to create ID for number '{$raw}'");
     }
 }