+++ src/NullIdFactory.php
@@ -7,13 +7,7 @@ namespace byrokrat\id;
  */
 class NullIdFactory extends IdFactory
 {
-    /**
-     * Create ID object
-     *
-     * @param  string $rawId Ignored for NullId
-     * @return IdInterface
-     */
-    public function create($rawId)
+    public function createId($raw)
     {
         return new NullId;
     }