+++ tests/IdFactoryTest.php
@@ -6,46 +6,41 @@ class IdFactoryTest extends \PHPUnit_Framework_TestCase
 {
     public function testPersonalIdFactory()
     {
-        $factory = new PersonalIdFactory();
         $this->assertInstanceOf(
-            '\byrokrat\id\PersonalId',
-            $factory->create('820323-2775')
+            PersonalId::CLASS,
+            (new PersonalIdFactory)->createId('820323-2775')
         );
     }
 
     public function testCoordinationIdFactory()
     {
-        $factory = new CoordinationIdFactory();
         $this->assertInstanceOf(
-            '\byrokrat\id\CoordinationId',
-            $factory->create('701063-2391')
+            CoordinationId::CLASS,
+            (new CoordinationIdFactory)->createId('701063-2391')
         );
     }
 
     public function testOrganizationIdFactory()
     {
-        $factory = new OrganizationIdFactory();
         $this->assertInstanceOf(
-            '\byrokrat\id\OrganizationId',
-            $factory->create('702001-7781')
+            OrganizationId::CLASS,
+            (new OrganizationIdFactory)->createId('702001-7781')
         );
     }
 
     public function testFakeIdFactory()
     {
-        $factory = new FakeIdFactory();
         $this->assertInstanceOf(
-            '\byrokrat\id\FakeId',
-            $factory->create('701023-xxxx')
+            FakeId::CLASS,
+            (new FakeIdFactory)->createId('701023-xxxx')
         );
     }
 
     public function testNullIdFactory()
     {
-        $factory = new NullIdFactory();
         $this->assertInstanceOf(
-            '\byrokrat\id\NullId',
-            $factory->create('')
+            NullId::CLASS,
+            (new NullIdFactory)->createId('')
         );
     }
 
@@ -60,6 +55,6 @@ class IdFactoryTest extends \PHPUnit_Framework_TestCase
         );
 
         $this->setExpectedException('byrokrat\id\Exception');
-        $factory->create('unvalid id');
+        $factory->createId('unvalid id');
     }
 }