+++ CHANGES.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Deprecated
 - The `Id` interface. Use `IdInterface` instead
+- `IdFactory::create()`. Use `createId()` instead.
 
 ## [1.0.2] - 2016-08-18